src/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
beautifulsoup4==4.14.2
boto3==1.40.54
botocore==1.40.54
certifi==2025.10.5
charset-normalizer==3.4.4
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.0.1
ipython==8.37.0
jedi==0.19.2
jiter==0.11.0
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.6
oauthlib==3.3.1
openai==2.3.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
text-unidecode==1.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
